+++ edc-extensions/hashicorp-vault/src/main/java/org/eclipse/tractusx/edc/hashicorpvault/HashicorpVaultClient.java
@@ -162,6 +162,9 @@ class HashicorpVaultClient {
   private HttpUrl getSecretUrl(String key, String entryType) {
     key = URLEncoder.encode(key, StandardCharsets.UTF_8);
 
+    // restore '/' characters to allow sub-directories
+    key = key.replace("%2F", "/");
+
     final String vaultApiPath = config.getVaultApiSecretPath();
 
     return Objects.requireNonNull(HttpUrl.parse(config.getVaultUrl()))