+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Set EDC version to `0.0.1-20221006-SNAPSHOT`
 -   Business Partner Number Extension no longer supports the 'IN' constraint operator
+-   HashiCorp Vault Extension now allows sub directories for secrets
 
 ### Changed